refactor: user-configを分割
